+++ NEWS.md
@@ -1,6 +1,6 @@
 # hmstimer 0.1.0
 
-- Deprecated `start` argument to `tmr_reset()`.
+- Hard deprecated `start` argument to `tmr_reset()`.
 - Added `tmr_format()`.
 - Added `tmr_round()`, `tmr_floor()` and `tmr_ceiling()`.
 - Switched to tidyverse style error and warning messages.